split class.
